requirements.txt
# Web scraper dependencies
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0

# Flask backend dependencies
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.2.1

# PostgreSQL dependencies
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
